refactor: search in platform adapter
EX-6313
