--- packages/x-adapter-platform/src/mappers/__tests__/response/search-response.mapper.spec.ts
@@ -1,12 +1,0 @@
-import { searchResponseMapper } from '../../responses/search-response.mapper';
-import { PlatformSearchResponse } from '../../../types/response.types';
-import { platformSearchResponse } from '../../../__tests__/__fixtures__/platform-search.response';
-import { searchResponse } from '../../../__tests__/__fixtures__/search.response';
-
-describe('search response mapper tests', () => {
-  it('should map the response', () => {
-    expect(
-      searchResponseMapper(platformSearchResponse as PlatformSearchResponse, {})
-    ).toStrictEqual(searchResponse);
-  });
-});
