+++ packages/x-adapter-platform/src/mappers/responses/search-response.mapper.ts
@@ -1,40 +1,8 @@
-import { createMutableSchema, Schema, schemaMapperFactory } from '@empathyco/x-adapter-next';
-import { getTaggingInfoFromUrl } from '@empathyco/x-utils';
-import { PlatformSearchResponse, SearchResponse } from '../../types/response.types';
-import { resultSchema } from '../../schemas/models/result.schema';
-import { facetMutableSchema } from '../../schemas/facet.schema';
-import { promotedMutableSchema } from '../../schemas/promoted.schema';
-import { bannerMutableSchema } from '../../schemas/banner.schema';
-import { redirectionMutableSchema } from '../../schemas/redirection.schema';
-
-export const searchResponseSchema: Schema<PlatformSearchResponse, SearchResponse> = {
-  results: {
-    $path: 'catalog.content',
-    $subSchema: resultSchema
-  },
-  facets: {
-    $path: 'catalog.facets',
-    $subSchema: facetMutableSchema
-  },
-  totalResults: 'catalog.numFound',
-  spellcheck: 'catalog.spellchecked',
-  banners: {
-    $path: 'banner.content',
-    $subSchema: bannerMutableSchema
-  },
-  promoted: {
-    $path: 'promoted.content',
-    $subSchema: promotedMutableSchema
-  },
-  redirections: {
-    $path: 'direct.content',
-    $subSchema: redirectionMutableSchema
-  },
-  queryTagging: ({ catalog }) => getTaggingInfoFromUrl(catalog?.tagging?.query)
-};
-
-export const mutableSearchResponseSchema = createMutableSchema(searchResponseSchema);
+import { schemaMapperFactory } from '@empathyco/x-adapter-next';
+import { SearchResponse } from '@empathyco/x-types';
+import { PlatformSearchResponse } from '../../types/responses/search-response.model';
+import { searchResponseMutableSchema } from '../../schemas/responses/search-response.schema';
 
 export const searchResponseMapper = schemaMapperFactory<PlatformSearchResponse, SearchResponse>(
-  mutableSearchResponseSchema
+  searchResponseMutableSchema
 );
