+++ packages/x-adapter-platform/src/schemas/filters/number-filter.schema.ts
@@ -1,8 +1,10 @@
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { NumberRangeFilter } from '@empathyco/x-types';
-import { PlatformFacetFilter } from '../../types';
+import { PlatformFilter } from '../../types/models/facet.model';
 
-export const numberFilterSchema: Schema<PlatformFacetFilter, NumberRangeFilter> = {
+export const numberFilterMutableSchema = createMutableSchema<
+  Schema<PlatformFilter, NumberRangeFilter>
+>({
   id: 'filter',
   facetId: 'id',
   label: 'value',
@@ -13,6 +15,4 @@ export const numberFilterSchema: Schema<PlatformFacetFilter, NumberRangeFilter> 
     min: ({ value }) => Number(value.split('-')[0]),
     max: ({ value }) => Number(value.split('-')[1])
   }
-};
-
-export const numberFilterMutableSchema = createMutableSchema(numberFilterSchema);
+});
