+++ packages/x-adapter-platform/src/schemas/filters/simple-filter.schema.ts
@@ -1,14 +1,12 @@
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { SimpleFilter } from '@empathyco/x-types';
-import { PlatformFacetFilter } from '../../types';
+import { PlatformFilter } from '../../types/models/facet.model';
 
-export const simpleFilterSchema: Schema<PlatformFacetFilter, SimpleFilter> = {
+export const simpleMutableFilterSchema = createMutableSchema<Schema<PlatformFilter, SimpleFilter>>({
   facetId: 'filter',
   label: 'value',
   id: 'filter',
   totalResults: 'count',
   selected: () => false,
   modelName: () => 'SimpleFilter'
-};
-
-export const simpleMutableFilterSchema = createMutableSchema(simpleFilterSchema);
+});
