+++ packages/x-adapter-platform/src/schemas/index.ts
@@ -2,8 +2,4 @@ export * from './filters';
 export * from './models';
 export * from './requests';
 export * from './responses';
-export * from './banner.schema';
-export * from './facet.schema';
-export * from './promoted.schema';
-export * from './redirection.schema';
 export * from './related-tag.schema';
