+++ packages/x-adapter-platform/src/schemas/models/banner.schema.ts
@@ -1,9 +1,9 @@
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { Banner } from '@empathyco/x-types';
 import { getTaggingInfoFromUrl } from '@empathyco/x-utils';
-import { PlatformBanner } from '../types/models.types';
+import { PlatformBanner } from '../../types/models/banner.model';
 
-export const bannerSchema: Schema<PlatformBanner, Banner> = {
+export const bannerMutableSchema = createMutableSchema<Schema<PlatformBanner, Banner>>({
   id: 'id',
   title: 'title',
   url: 'url',
@@ -12,6 +12,4 @@ export const bannerSchema: Schema<PlatformBanner, Banner> = {
   tagging: {
     query: ({ tagging }) => getTaggingInfoFromUrl(tagging?.query ?? '')
   }
-};
-
-export const bannerMutableSchema = createMutableSchema(bannerSchema);
+});
