+++ packages/x-adapter-platform/src/schemas/models/index.ts
@@ -1,3 +1,7 @@
-export * from './suggestion.schema';
+export * from './banner.schema';
+export * from './facet.schema';
 export * from './next-query.schema';
+export * from './promoted.schema';
+export * from './redirection.schema';
 export * from './result.schema';
+export * from './suggestion.schema';
