+++ packages/x-adapter-platform/src/schemas/models/promoted.schema.ts
@@ -1,9 +1,9 @@
 import { Promoted } from '@empathyco/x-types';
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { getTaggingInfoFromUrl } from '@empathyco/x-utils';
-import { PlatformPromoted } from '../types/models.types';
+import { PlatformPromoted } from '../../types/models/promoted.model';
 
-export const promotedSchema: Schema<PlatformPromoted, Promoted> = {
+export const promotedMutableSchema = createMutableSchema<Schema<PlatformPromoted, Promoted>>({
   id: 'id',
   url: 'url',
   title: 'title',
@@ -12,6 +12,4 @@ export const promotedSchema: Schema<PlatformPromoted, Promoted> = {
   tagging: {
     query: ({ tagging }) => getTaggingInfoFromUrl(tagging?.query ?? '')
   }
-};
-
-export const promotedMutableSchema = createMutableSchema(promotedSchema);
+});
