+++ packages/x-adapter-platform/src/schemas/models/redirection.schema.ts
@@ -1,15 +1,15 @@
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { Redirection } from '@empathyco/x-types';
 import { getTaggingInfoFromUrl } from '@empathyco/x-utils';
-import { PlatformRedirection } from '../types/models.types';
+import { PlatformRedirection } from '../../types/models/redirection.model';
 
-const redirectionSchema: Schema<PlatformRedirection, Redirection> = {
+export const redirectionMutableSchema = createMutableSchema<
+  Schema<PlatformRedirection, Redirection>
+>({
   id: 'id',
   url: 'url',
   modelName: () => 'Redirection',
   tagging: {
     click: ({ tagging }) => getTaggingInfoFromUrl(tagging?.click ?? '')
   }
-};
-
-export const redirectionMutableSchema = createMutableSchema(redirectionSchema);
+});
