+++ packages/x-adapter-platform/src/schemas/requests/index.ts
@@ -1,4 +1,5 @@
+export * from './identifier-results-request.schema';
+export * from './next-queries-request.schema';
 export * from './popular-searches-request.schema';
 export * from './query-suggestions.schema';
-export * from './next-queries-request.schema';
-export * from './identifier-results-request.schema';
+export * from './search-request.schema';
