+++ packages/x-adapter-platform/src/schemas/responses/index.ts
@@ -1,4 +1,5 @@
+export * from './identifier-results-response.schema';
+export * from './next-queries-response.schema';
 export * from './popular-searches-response.schema';
 export * from './query-suggestions-response.schema';
-export * from './next-queries-response.schema';
-export * from './identifier-results-response.schema';
+export * from './search-response.schema';
