+++ packages/x-adapter-platform/src/types/models/index.ts
@@ -1,4 +1,8 @@
-export * from './suggestion.model';
+export * from './banner.model';
+export * from './facet.model';
 export * from './next-query.model';
+export * from './promoted.model';
+export * from './redirection.model';
 export * from './result.model';
+export * from './suggestion.model';
 export * from './tagging.model';
