+++ packages/x-adapter-platform/src/types/models/promoted.model.ts
@@ -1,0 +1,14 @@
+/**
+ * Promoted model for the `platform` API.
+ *
+ * @public
+ */
+export interface PlatformPromoted {
+  id: string;
+  title: string;
+  url: string;
+  image_url: string;
+  tagging?: {
+    query: string;
+  };
+}
