+++ packages/x-adapter-platform/src/types/models/redirection.model.ts
@@ -1,0 +1,12 @@
+/**
+ * Redirection model for the `platform` API.
+ *
+ * @public
+ */
+export interface PlatformRedirection {
+  id: string;
+  url: string;
+  tagging?: {
+    click: string;
+  };
+}
