+++ packages/x-adapter-platform/src/types/request.types.ts
@@ -1,6 +1,3 @@
-import { Filter } from '@empathyco/x-types';
-import { Dictionary } from '@empathyco/x-utils';
-
 export interface BaseRequest {
   rows: number;
   start: number;
@@ -22,22 +19,6 @@ export interface PlatformBaseRequest {
   lang: string;
   query?: string;
   origin?: string;
-}
-
-export interface SearchRequest extends BaseRequest {
-  filters: Dictionary<Filter[]>;
-  query: string;
-  origin: string;
-  relatedTags?: any[];
-  sort: string;
-  lang: string;
-}
-
-export interface PlatformSearchRequest extends PlatformBaseRequest {
-  origin: string;
-  query: string;
-  filter?: string[];
-  sort?: string;
 }
 
 export interface TaggingRequest {
