+++ packages/x-adapter-platform/src/types/response.types.ts
@@ -1,41 +1,7 @@
-import {
-  Banner,
-  NextQuery,
-  Promoted,
-  Redirection,
-  RelatedTag,
-  Result,
-  TaggingRequest
-} from '@empathyco/x-types';
-
-import {
-  Contentable,
-  PlatformBanner,
-  PlatformCatalog,
-  PlatformPromoted,
-  PlatformRedirection,
-  PlatformRelatedTag
-} from './models.types';
+import { NextQuery, RelatedTag, Result } from '@empathyco/x-types';
+import { PlatformRelatedTag } from './models.types';
 import { PlatformResult } from './models/result.model';
 import { PlatformNextQuery } from './models/next-query.model';
-
-export interface PlatformSearchResponse {
-  banner: Contentable<PlatformBanner>;
-  catalog: PlatformCatalog;
-  direct: Contentable<PlatformRedirection>;
-  promoted: Contentable<PlatformPromoted>;
-}
-
-export interface SearchResponse {
-  results: Result[];
-  facets: any[];
-  totalResults: number;
-  spellcheck: string;
-  banners: Banner[];
-  promoted: Promoted[];
-  redirections: Redirection[];
-  queryTagging: TaggingRequest;
-}
 
 export interface PlatformTopClickedResponse {
   topclicked: {
