+++ packages/x-adapter-platform/src/types/responses/index.ts
@@ -1,3 +1,4 @@
 export * from './identifier-results-response.model';
 export * from './query-suggestions-response.model';
 export * from './popular-searches-response.model';
+export * from './search-response.model';
